ts7180: Honor U-Boot jumper & add push-switch snapshot (jpsw) to env
As part of this, I removed legacy/ported code that lets the
push-switch act like the u-boot jumper. It would interfere with
wake-from-sleep, which *needs* auto-boot in order to function
correctly.

Prior behavior can be approximated by modifying the 'bootcmd'
environment variable to read the value of 'jpsw'. jpsw is the snapshot
value of the push switch at the time of U-Boot initialization.