technologic: ts7250v3: Added support for REV C PCBs
This adds support for the REV C which prints the microcontroller revision, reboot reason, and
supports a new method for poweroff and reset.  The defconfig is updated to not print the
imx6ul reset reason, and we print ours in the board instead. checkboard was moved to run
in late boot rather than early boot.