ts7100: WDOG reset used by 5.10 kernel alters FPGA startup timing
After a CPU WDOG# reset (as opposed to a Silabs watchdog reset), the
DRAM sizing strap on the FPGA reads as 1GB for 512MB boards.

Reads OK with 4.9 and U-Boot resets, as well as from power-on.

This commit is associated with FPGA rev 23. It uses the FPGA_INT2 GPIO
to signal the CPU when it is finished with its reset. Up to 45 ms
will be given for the FPGA to drive FPGA_INT2 low.

As of this commit, only 5 mdelay calls are needed before the FPGA
reports sane and gives the correct DRAM strapping.