drivers: imx_rgpio2p: Use device tree alias for bank numbers
The existing behavior uses:
GPIO1_
GPIO2_
...

This matches the GPIO numbers in the datasheet. Except, when adding additional GPIO,
eg 3 banks off of flexspi, it bumps these to:
GPIO4_
GPIO5_
...

Instead of using the ID from the number of GPIO drivers loaded, use the device tree aliases so
these can be consistent.